Correctly detect functions that belong to a receiver
Previously functions that had a name starting with "New" followed by the
name of a type that was part of its return value was considered a
constructor. The way we handle it now is more like the way go doc
handles these cases. A function belongs to a receiver based on its first
return type. If it's a named return type, it must not have multiple
names. In this case a function is not a constructor any more, but a
function with its receiver set.