Correctly handle multiple named return types
We do not use named parameters in return types for a tag. However, when
the return types are named and there are multiple names for a type it
would only show that type once in the tag.

For example, the following function

	func t() (a, b string, e error)

needs to create these return types in a tag

	string, string, error

Before this commit it would only create

	string, error